daemon: improve handling stalling VM's daemon
qubesdb-daemon on the other side of the vchan connection could write
partial message and then stall. Previously this could cause dom0's
qubesdb-daemon instance to stall in processing other requests (including
from qubesd) too. Rework vchan requests handling to avoid this.

Specifically: retrieve message header only if the whole header is there.
And then check if the whole message data is already available on vchan.
If not - cache the header and return to the mainloop for other events
processing, until remaining part of the message body is sent.

This should be resiliant against broken qubesdb instance, but not
necessary malicious. This is because there is TOCTOU issue in vchan
itself - malicious VM could alter vchan metadata between check for
available data and actually retrieving it. Fix for this requires
setting "blocking" flag in vchan to false, which isn't exposed in public
API yet.

QubesOS/qubes-issues#5073